+++ changelog.md
@@ -107,3 +107,10 @@
     * Renamed delete_machine to machine_delete
     * Renamed delete_machines to machines_delete
     * Added get_application_pools
+
+22-02-2021
+* Inventory Class
+    * Added get_application_pool
+    * Added new_application_pool
+    * Added update_application_pool
+    * Renamed machine related methods